Refactored vcs code, finally.
